requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-google-genai
pydantic
python-dotenv